+++ cops
@@ -1,0 +1,1 @@
+rubocop/cops/
